+++ src/Equinor.Procosys.Preservation.WebApi/Controllers/Journey/JourneysController.cs
@@ -5,7 +5,7 @@ using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 
-namespace Equinor.Procosys.Preservation.WebApi.Controllers
+namespace Equinor.Procosys.Preservation.WebApi.Controllers.Journey
 {
     [ApiController]
     [Route("Journeys")]
