+++ src/Equinor.Procosys.Preservation.WebApi/Controllers/Mode/ModesController.cs
@@ -7,7 +7,7 @@ using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 
-namespace Equinor.Procosys.Preservation.WebApi.Controllers
+namespace Equinor.Procosys.Preservation.WebApi.Controllers.Mode
 {
     [ApiController]
     [Route("Modes")]
