+++ src/Equinor.Procosys.Preservation.WebApi/Controllers/Tag/CreateTagDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Equinor.Procosys.Preservation.Command.TagCommands
+﻿namespace Equinor.Procosys.Preservation.WebApi.Controllers.Tag
 {
     public class CreateTagDto
     {
